git-svn-id: http://10.0.0.3/svn/poker@361 f0252c50-933b-4dd8-9fc7-05cab60132cc

+++ database.py
@@ -166,7 +166,8 @@ def init_database():
 
 if __name__ == "__main__":
 	import os
-	os.remove("db.sqlite")
+	if os.path.isfile("./db.sqlite"):
+		os.remove("db.sqlite")
 	db_connection  = DatabaseConnection()
 	db_connection.init("sqlite:///db.sqlite")
 	db_connection.connect()
